- moved PAS installation to setuphandlers.py - fixes
