+++ src/plone/webauthn/configure.zcml
@@ -18,8 +18,8 @@
   <include file="permissions.zcml" />
 
   <adapter
-      factory="plone.webauthn.src.plone.webauthn.key_data.KeyDataAdapter"
-      provides="plone.webauthn.src.plone.webauthn.key_data.IKeyData"
+      factory="plone.webauthn.key_data.KeyDataAdapter"
+      provides="plone.webauthn.key_data.IKeyData"
       for="*"
     />
