+++ test_plone60.cfg
@@ -59,3 +59,12 @@ loguru = 0.7.0
 # Required by:
 # plone.webauthn==1.0a1
 zopyx.plone.persistentlogger = 0.5.1
+
+# Added by buildout at 2023-06-26 18:28:28.064398
+asn1crypto = 1.5.1
+cbor2 = 5.4.6
+pydantic = 1.10.9
+
+# Required by:
+# plone.webauthn==1.0a1
+webauthn = 1.8.1
